Rename *.common test files to *.playback
